requirements.txt
aiohttp==3.8.4
APScheduler==3.10.1
beautifulsoup4==4.12.2
dataframe_image==0.1.11
discord.py==2.3.1
gql==3.4.1
matplotlib==3.7.1
numpy==1.25.0
pandas==2.0.3
python-dotenv==1.0.0
backoff==2.2.1
Requests==2.31.0
web3==6.20.1